+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 - feat(services): add a mechanism to run external processes as part of ubo services, these processes are supposed to communicate to the ubo-app via gRPC, they have their own installation script (`ubo-setup.sh` file in the service directory) and run in a separate virtualenv
 - refactor(rpc): several improvements in rpc and store types, most importantly, avoid marking all fields as optional in proto files and remove `items` from `UboPageWidget`'s constructor
 - refactor(audio): encapsulate audio sample data in the new `AudioSample` data structure, better support playing streaming sequence of audio sample by guaranteeing sequential playback and reporting when playback is done
+- feat(assistant): add pipecat in a separate process
 
 ## Version 1.4.0